Testing interface

+++ database/migrations/2020_11_28_135203_products.php
@@ -16,7 +16,7 @@ class Products extends Migration
        Schema::create('products', function (Blueprint $table) {
            $table->id();
            $table->string('name');
-           $table->float('price');
+           $table->decimal('price');
            $table->timestamps();
        });
    }
